refactor: address crosscheck Notable Ambiguous Q-a..Q-g on PR #740
Five user-direction decisions on crosscheck Notable Ambiguous items:

Q-a — RefRole / RefIn* naming divergence from sibling ChildRole / Role*:
  Decision: leave as-is. The Ref/Role split is semantically real
  (Ref = 'what this use IS'; Role = 'what slot a child fills'); the
  cosmetic similarity does not justify a rename across the package.

Q-b — visitRecord exported fields (Name/Role/Depth):
  Decision: keep exported. Table-driven qt.DeepEquals beats per-field
  c.Assert for readability; exported-on-unexported-types is a known
  Go idiom signaling 'for reflection only'.

Q-c — qt.Assert vs c.Assert style split in internal/validate/ tests:
  Decision: leave mixed, file a TODO. The new file matches the most-
  recent precedent (env_helpers_test.go from PR #739). Package-wide
  alignment is a separate sweep recorded in TODO.md under
  'Internal-SR follow-ups (PR #739 deferred items)'.

Q-d — Move buildBindingIdxMap to env_helpers.go:
  Decision: move. The helper matches the verb-noun env-resolution
  family (bindLocalSymbol, extendEnvWithSymbols, findDuplicateSymbols).
  Renamed nothing — the name was already verb-noun-shaped. The
  walk_binding_refs.go forward-comment preserves discoverability.

Q-e — Exhaustive switch on RefRole at consumer call sites:
  Decision: apply to escape only. The capture visitor filters on
  depth (role ignored); converting it to a switch would add ceremony
  with no invariant gain. The escape visitor filters on role; converted
  to an exhaustive switch with cases RefInBody / RefInCallProc /
  RefSetBangTarget / default-panic. A future 4th role now surfaces
  at the consumer call site rather than silently being classified as
  'not escaping'.

Q-f — Add ValidatedDefine depth-tagging tests:
  Decision: add two. TestWalkBindingRefs_DefineFunctionDepth pins
  '(define (f) x)' yielding x at depth=1 (body crosses closure
  boundary). TestWalkBindingRefs_DefineValueSameDepth pins
  '(define y x)' yielding x at depth=0 (value form, no closure).
  Pins a contract currently covered only transitively through
  capture/escape integration tests.

Q-g — Compound visit-order test:
  Decision: add. TestWalkBindingRefs_CallProcInsideClosureBody walks
  '(lambda () (g x))' and asserts g at RefInCallProc/depth=1, x at
  RefInBody/depth=1. Pins the composition of call-proc tagging with
  depth increment AND the operator-before-arg visit order inside a
  closure body.

3 new unit tests added (15 total in walk_binding_refs_test.go).
All existing tests still pass. Build + lint + covercheck + ci clean.

Refs: PR #740; crosscheck output (5 lenses)
